+++ src/Repository/UserRepository.php
@@ -2,19 +2,16 @@
 
 namespace App\Repository;
 
-use App\Entity\Societe;
 use App\Entity\User;
-use App\Exception\RdiException;
-use App\HasSocieteInterface;
-use App\Role;
-use App\Security\Role\RoleProjet;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\ORM\QueryBuilder;
 use Doctrine\Persistence\ManagerRegistry;
+use libphonenumber\PhoneNumber;
 
 /**
  * @method User|null find($id, $lockMode = null, $lockVersion = null)
  * @method User|null findOneBy(array $criteria, array $orderBy = null)
+ * @method User|null findOneByEmail(string $email)
+ * @method User|null findOneByTelephone(PhoneNumber $telephone)
  * @method User[]    findAll()
  * @method User[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
  */