+++ src/Service/RdiMailer.php
@@ -3,12 +3,9 @@
 namespace App\Service;
 
 use App\DTO\RecommandationMessage;
-use App\Entity\User;
-use App\Exception\RdiException;
 use Symfony\Bridge\Twig\Mime\TemplatedEmail;
 use Symfony\Component\Mailer\MailerInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
-use Symfony\Contracts\Translation\TranslatorInterface;
 
 class RdiMailer
 {
@@ -16,19 +13,15 @@ class RdiMailer
 
     private $mailer;
 
-    private $translator;
-
     private $urlGenerator;
 
     public function __construct(
         string $mailFrom,
         MailerInterface $mailer,
-        TranslatorInterface $translator,
         UrlGeneratorInterface $urlGenerator
     ) {
         $this->mailFrom = $mailFrom;
         $this->mailer = $mailer;
-        $this->translator = $translator;
         $this->urlGenerator = $urlGenerator;
     }
 
@@ -76,34 +69,6 @@ class RdiMailer
             ->textTemplate('mail/_recommandation-content.txt.twig')
 
             ->htmlTemplate('mail/recommandation.html.twig')
-        ;
-
-        $this->mailer->send($email);
-    }
-
-    public function sendResetPasswordEmail(User $user): void
-    {
-        if (!$user->hasResetPasswordToken()) {
-            throw new RdiException('Cannot send reset password email, this user has no reset password token.');
-        }
-
-        $resetPasswordLink = $this->urlGenerator->generate('app_fo_reset_password', [
-            'token' => $user->getResetPasswordToken(),
-        ], UrlGeneratorInterface::ABSOLUTE_URL);
-
-        $email = $this->createDefaultEmail()
-            ->to($user->getEmail())
-            ->subject(sprintf('Réinitialisation de votre mot de passe RDI-Manager'))
-            ->text(sprintf(
-                'Vous avez demandé un lien de réinitialisation de votre mot de passe. '
-                .'Suivez ce lien pour définir votre nouveau mot de passe : %s',
-                $resetPasswordLink
-            ))
-
-            ->htmlTemplate('mail/reset_password.html.twig')
-            ->context([
-                'resetPasswordLink' => $resetPasswordLink,
-            ])
         ;
 
         $this->mailer->send($email);