+++ src/Service/ResetPasswordService.php
@@ -4,7 +4,11 @@ namespace App\Service;
 
 use App\Entity\User;
 use App\Exception\ResetPasswordException;
+use App\Notification\Event\ResetPasswordRequestNotification;
 use App\Repository\UserRepository;
+use libphonenumber\NumberParseException;
+use libphonenumber\PhoneNumberUtil;
+use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
 class ResetPasswordService
 {
@@ -12,34 +16,53 @@ class ResetPasswordService
 
     private UserRepository $userRepository;
 
-    private RdiMailer $mailer;
+    private PhoneNumberUtil $phoneNumberUtil;
+
+    private EventDispatcherInterface $dispatcher;
 
     public function __construct(
         TokenGenerator $tokenGenerator,
         UserRepository $userRepository,
-        RdiMailer $mailer
+        PhoneNumberUtil $phoneNumberUtil,
+        EventDispatcherInterface $dispatcher
     ) {
         $this->tokenGenerator = $tokenGenerator;
         $this->userRepository = $userRepository;
-        $this->mailer = $mailer;
+        $this->phoneNumberUtil = $phoneNumberUtil;
+        $this->dispatcher = $dispatcher;
     }
 
     /**
-     * @param string $email De l'utilisateur qui fait la demande.
+     * @param string $username Email ou SMS de l'utilisateur.
      *
      * @return User Qui a fait la demande de réinitialisation, et qui a maintenant un token valide.
      *
      * @throws ResetPasswordException Si l'email n'est pas attribué.
      */
-    public function requestLink(string $email): User
+    public function requestLink(string $username): User
     {
-        $user = $this->userRepository->findOneBy([
-            'email' => $email,
-            'enabled' => true,
-        ]);
+        if (str_contains($username, '@')) {
+            $user = $this->userRepository->findOneBy([
+                'email' => $username,
+            ]);
+        } else {
+            try {
+                $phoneNumber = $this->phoneNumberUtil->parse($username, 'FR');
+            } catch (NumberParseException $e) {
+                throw new ResetPasswordException('Ce numéro de téléphone semble incorrect.', $e);
+            }
+
+            $user = $this->userRepository->findOneBy([
+                'telephone' => $phoneNumber,
+            ]);
+        }
 
         if (null === $user) {
-            throw new ResetPasswordException('Cet email n\'est attribué à aucun utilisateur.');
+            throw new ResetPasswordException('Cet email ou n° de téléphone n\'est attribué à aucun utilisateur.');
+        }
+
+        if (!$user->getEnabled()) {
+            throw new ResetPasswordException('Ce compte est désactivé.');
         }
 
         $token = $this->tokenGenerator->generateUrlToken();
@@ -50,7 +73,7 @@ class ResetPasswordService
             ->setResetPasswordTokenExpiresAt($expirationDate)
         ;
 
-        $this->mailer->sendResetPasswordEmail($user);
+        $this->dispatcher->dispatch(new ResetPasswordRequestNotification($user));
 
         return $user;
     }